20240620/2조/code/requirements.txt
torch>=1.13.1
torchvision
numpy
Pillow
matplotlib
scipy